requirements.txt
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
streamlit-aggrid==1.0.5 